Issue stacktrace: Don't show variables on the RHS
in the end I find this harder to read. better to keep them in the same place always.
making this move prepares us for the 'key issue insights' box too, because it provides
insight in the 'amount of space used' by the main tab (for which the code lines in the
stacktrace will be the largest)