Show 'logger' if this info is available
Somewhat untested, we used to have this in a very early version ("Emitted by")
and it was removed when converting into the calculated_xx style. I haven't found
any recent sample events that have this value but according to the spec it can
exist so I'm putting it back in.