Don't call the view yourself (and return a result) from process_view
calling the view yourself short-circuits the correct handling, which is not
what we want. It's not abundantly clear from the docs that this is generally
not what we want, but it is cimplied:

> It should return either None or an HttpResponse object. If it returns None,
> Django will continue processing this request, executing any other
> process_view() middleware and, then, the appropriate view. If it returns an
> HttpResponse object, Django won’t bother calling the appropriate view; it’ll
> apply response middleware to that HttpResponse and return the result.

We don't want to do any short-circuiting here, we just want to measure time
and have a view_name to display. Moving stuff around fixes it.

(That this was an error became clear when I wanted to do process_exception
in another Middleware, and that was never called)