Remove auto_session_tracking=False from sentry-init
The default is True.

It was copy-pasted from earlier integrations, but its exact origin (and
whether it is strictly needed) is unclear. Perhaps installation instructions
from (some) Sentry or Glitchtip?

We don't have explicit support for sessions, but I'm not clear enough on this
one to recommend against it either. And for our own usage, this provides a
way to get samples of session data in the playground/dogfood.