SECRET_KEY: make it required
Considered: having a fall-back that is generated on-server-start. That doesn't
fly because we have gunicorn, and each server gets their own value.

Generating once on first run (in the Docker container), writing to a file: way
too fragile. i.e. on-container-restart you'd get the same (and this would be
surprising).

Given the sensitive nature of all of this, I'd say being explicit

The actual effect of changes is limited to sessions invalidating (but that bad
enough).

And the current setup is more broad, since it does a general check.

See https://stackoverflow.com/a/30266422/339144 (note: PasswordResetView
isn't used by us, we have our own)