Remove another comment about ALLOWED_HOSTS
We already removed this one from the place where it matters most, i.e. the
recommended template.

The key bit of info is: the hard bit is in doing this for HTTPS/SSL, because
in that case you don't have a certificate (it could be for any web address).
Details are in the website's repo, in templates/production-install.html.