Remove dashses from displayed DSN
Fixes #7

I could not find any documentation on what the "standard" is, but I know that
we'll pick it up just fine either way (because Django's UUID field does the
magic for us).

Given that it's impossible to setup your JS client with the dashes, they should
simply be removed.