deduce_allowed_hosts: localhost/127.0.0.1 -> *
Reverts part of 8ff309cd8f7a; in that commit I was inspired by the comment
right above it to make it more strict, but taking a step back (and inspired
by a user tripping over ALLOWED_HOSTS validation) defaulting to "*" for
basically non-configured setups is actually better.

Preserved from 8ff309cd8f7a: the idea that both 127.0.0.1 and localhost are
seen as localhost.