Attempt at better guess_lexer_for_filename
the idea: guess on filename first, and on code second.
prompted by: getting a Django Template Parser to work.

however, this won't fly, because lexer.analyse_text is basically useless
(the parts that are implemented __at all__ are often based on some initial
bytes in the file, which is useless in our context of code snippets)