In bad_request handler, don't refetch the exception
Not prompted by an observed bug, but by the observation that exception is already
passed in as an argument. Likely, this was copied from the '500' case, where it
is not. Despite not having observed a bug, the improved version is more likely
to be correct (I don't want to have to think about how sys.exc_info() interacts
with the template-failure-handling right above)