PoC for file storage; using the on-commit approach which I'll probably abandon
This branch shows an approach similar to 5559fba7542c32f78aa65dd54f6edfedd31ce963,
but with writing and deleting to the filestore pulled out of the transaction.

The idea would be for slower file backends (e.g. S3, over the network). However:

however, the `immediate_semaphore` that I built in is actually only released after
the `on_commit` functions are run. So in the plain case this simply doesn't work.
In particular, super call to `Atomic` triggers the following before the release happens:

* `connection.set_autocommit(True)` (in `django/db/transaction.py`)
* `run_and_clear_commit_hooks` (in `django/db/backends/base/base.py`)

Some ways forward (once S3 or similar are introduced):

* Add your own `on_xx`, which you can call after release of the semaphore.
    (this is the preferred approach, I think!)

* copy/paste the whole of `Atomic.__exit__` and just blunt-axe it in.
    in this approach the problem that `immeadiate_semaphore` solves must be
    re-established (tests?) before I proceed though.

* Heavier snappea use (delaying, which will get you out of the transaction

A separate concern: heavy delay use, with use of (durable) transactions, and
EAGER mode is not a happy combination (the eagerness makes the transactions
nested)

Note that the "inside transaction approach" (the one from 5559fba7542c) is actually
_more correct_ for the local-FS; so doing it on-commit must perhaps be _only_ done
for slow stores? point-of-config on the storage class?

When taking a delayed approach (and when dealing with unreliable, i.e. networked,
storages), we also need more explicit handling of errors when the stored event is
missing (i.e. in the issue-views).

on-instance caching for already-read results is another todo for the S3 scenario.
i.e. event._cached_event_data