Remove 'order_by' from delete_with_deps logic
I _think_ I added this in the first version for [1] some idea around
determinism and [2] with the idea that by "grouping" around fk_name,
you'd get deletions close-to-root sooner, but I don't think either of
those are valuable goals (in both cases: as long as it's gone eventually).
The drawback of ordering is that it forces further thinking about indexes
(esp. when ordering on 2 items) so I'd rather avoid that.