Skip ALLOWED_HOSTS validation for /health/ endpoints
Health checks are often done directly against an IP address rather than using a
full hostname. This is annoying to set up if the IP address can be one from a
large range.

The things that ALLOWED_HOSTS defend against don't really apply for health
endpoints, so better not have the check.

Fix #140