Sourcemaps upload: max file size 2GiB
As per the comment.

Since we haven't actually gone multi-chunk, this is just preparation

The now-removed comment should be read as 'it could be assumed that
unzipping introduces a factor 5 increase between chunk size and file
size' but that's a whole bunch of assumptions that I'd rather get
rid of (mental overhead, with little gain).

See #147