File-upload: update comments about `concurrency` and `chunksPerRequest`
i.e. update the comments to reflect what I just learned doing some actual
experiments.

See #147

b.t.w. the now-removed comment was somewhat misleading: "single-chunk"
was (and is) being forced as in "single chunk per request" but not as
in "single chunk per file", and it was only forced by chunksPerRequest=1,
not by concurrency=1.