DJDT: pin at 5.2
'philosphofically' I prefer to keep my dev-deps in flux ('bleeding edge')
but since I barely use the DJDT I'd rather just pin it at a known-working
version.

Also: 6.0 introduces DB-models (for a debug tool) which I'm not a fan of.

Probably removing DJDT right after this, which would make this commit to
be a good point to revert to if we ever want to reintroduce it

See #168