yesno filter: just don't return None ever
I had someone run into this on hosted Bugsink; couldn't reproduce it.
Thought I fixed it in e8fb9556f723 (specific to the Chinese translation)
but appararently there's other ways to reach this point.

No matter, just create a version of the yesno filter that's not sensitive
to any future mistranslation.